# api/openapi.yaml
openapi: 3.0.3
info:
  title: Calendar API
  version: 1.0.0
  contact:
    name: Source code
    url: https://github.com/Alexander-D-Karpov/calendar
  description: |
    JSON API for calendars, events, todos and share links. All operations live under `/api/v1`.

    ## Authentication

    Send an API token in the `Authorization` header:

        Authorization: Bearer cal_<prefix>_<secret>

    Create tokens in [Settings → API tokens](/settings/tokens). A token is shown once, right after it is created, and can be revoked at any time. To try requests on this page with a token, press **Authorize** and paste it.

    Requests from a signed-in browser session work too. With a session, unsafe methods (`POST`, `PUT`, `PATCH` and `DELETE`) also need the `X-CSRF-Token` header. **Try it out** on this page adds that header when you are signed in.

    ## Scopes

    Each token carries one or more scopes, and each operation names the scope it needs. A `write` scope includes `read` access to the same resource. Browser sessions have every scope.

    | Scope | Allows |
    |---|---|
    | `account:read` | Reading the account profile |
    | `account:write` | Changing account settings and the sleep schedule |
    | `calendars:read` | Reading calendars and events |
    | `calendars:write` | Creating, changing and deleting calendars and events |
    | `todos:read` | Reading todo lists, todos and checks |
    | `todos:write` | Creating, changing, completing and deleting todos |
...

...

    ## MCP

...

...
servers:
...
paths:
...
  /calendars:
    get:
      tags: [Calendars]
      operationId: listCalendars
      summary: List calendars
      description: |
        Returns every calendar of the account ordered by `position`. Hidden calendars are included.

        Requires scope `calendars:read`.
      x-scope: calendars:read
      security: *readAuth
      responses:
        "200":
          description: All calendars of the account.
          headers: &okHeaders
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarList"
              example:
                items:
                  - &calendarExample
                    id: "0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10"
                    name: Work
                    color: "#3b6ea5"
                    description: ""
                    timezone: Europe/Moscow
                    kind: local
                    read_only: false
                    is_default: false
                    hidden: false
                    position: 1
                    default_reminders: [10]
                    created_at: "2026-09-10T13:56:42Z"
                    updated_at: "2026-09-10T14:02:11Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Calendars]
      operationId: createCalendar
      summary: Create a calendar
      description: |
        Creates a local calendar. Only `name` is required. The calendar is placed after the others unless `position` is set. An account can have at most 100 calendars.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarInput"
            example:
              name: Work
              color: "#3b6ea5"
              timezone: Europe/Moscow
              default_reminders: [10]
      responses:
        "201":
          description: The created calendar.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Location:
              $ref: "#/components/headers/Location"
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
              example: *calendarExample
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /calendars/{id}:
    parameters:
      - $ref: "#/components/parameters/CalendarID"
    get:
      tags: [Calendars]
      operationId: getCalendar
      summary: Get a calendar
      description: |
        Returns one calendar. Send the `ETag` from an earlier response in `If-None-Match` to get `304` when nothing changed.

        Requires scope `calendars:read`.
      x-scope: calendars:read
      security: *readAuth
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: The calendar.
          headers: *resourceHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
              example: *calendarExample
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    patch:
      tags: [Calendars]
      operationId: updateCalendar
      summary: Update a calendar
      description: |
        Changes the fields present in the body as a JSON Merge Patch. Send `"timezone": null` to fall back to the account timezone. `kind`, `read_only` and `is_default` cannot be changed.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/CalendarInput"
            example:
              name: Personal
              timezone: null
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarInput"
            example:
              hidden: true
      responses:
        "200":
          description: The updated calendar.
          headers: *resourceHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
              example: *calendarExample
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      tags: [Calendars]
      operationId: deleteCalendar
      summary: Delete a calendar
      description: |
        Deletes the calendar together with all of its events and removes it from share links. The default calendar cannot be deleted.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: The calendar and its events were deleted.
          headers: *okHeaders
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: cal_<prefix>_<secret>
      description: API token created in Settings → API tokens.
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-sid
      description: >-
        Browser session cookie, set by signing in. It is named `sid` when the server runs over plain HTTP.
        Unsafe methods also need the `X-CSRF-Token` header.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >-
        CSRF token of the browser session, required together with `sessionCookie` on unsafe methods.
        Pages expose it in the `csrf-token` meta tag.
  parameters:
    CalendarID:
      name: id
      in: path
      required: true
      description: Calendar ID.
      schema:
        type: string
        format: uuid
      example: "0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10"
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: ETag from an earlier response. The request fails with `412` if the resource changed since then.
      schema:
        type: string
      example: '"lx3kq9zb"'
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: ETag from an earlier response. The server answers `304` without a body if the resource is unchanged.
      schema:
        type: string
      example: '"lx3kq9zb"'
...
  headers:
    ETag:
      description: Version of the resource. Send it in `If-Match` to avoid overwriting other changes.
      schema:
        type: string
      example: '"lx3kq9zb"'
    Location:
      description: URL of the created resource.
      schema:
        type: string
      example: /api/v1/calendars/0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10
    RateLimitLimit:
      description: Requests allowed per window.
      schema:
        type: integer
        minimum: 1
      example: 600
    RateLimitRemaining:
      description: Requests left in the current window.
      schema:
        type: integer
        minimum: 0
      example: 599
    RateLimitReset:
      description: Seconds until the full quota is available again.
      schema:
        type: integer
        minimum: 0
      example: 1
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
        minimum: 1
      example: 12
    RequestID:
      description: Request identifier, also returned in error bodies as `request_id`.
      schema:
        type: string
      example: tQ6DNZRbQcyrEPPA
  responses:
...
    NotModified:
      description: The resource matches the ETag in `If-None-Match`.
      headers:
        ETag:
          $ref: "#/components/headers/ETag"
    Unauthorized:
      description: The token or session is missing, invalid, revoked or expired.
      headers:
        WWW-Authenticate:
          description: Authentication scheme to use.
          schema:
            type: string
          example: 'Bearer realm="calendar"'
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Unauthorized
            status: 401
            detail: invalid api token
            instance: /api/v1/me
            request_id: tQ6DNZRbQcyrEPPA
    Forbidden:
      description: The token lacks the required scope, or a session request failed the CSRF or origin check.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            scope:
              summary: Missing scope
              value:
                type: "about:blank"
                title: Forbidden
                status: 403
                detail: "token lacks scope calendars:write"
                instance: /api/v1/calendars
                request_id: tQ6DNZRbQcyrEPPA
            csrf:
              summary: Missing CSRF token
              value:
                type: "about:blank"
                title: Forbidden
                status: 403
                detail: invalid csrf token
                instance: /api/v1/calendars
                request_id: tQ6DNZRbQcyrEPPA
    NotFound:
      description: The resource does not exist or belongs to another account.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Not Found
            status: 404
            instance: /api/v1/calendars/0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10
            request_id: tQ6DNZRbQcyrEPPA
    Conflict:
      description: The request conflicts with the current state, for example deleting the default calendar.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Conflict
            status: 409
            detail: the default calendar cannot be deleted
            instance: /api/v1/calendars/0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10
            request_id: tQ6DNZRbQcyrEPPA
    PreconditionFailed:
      description: The resource changed since the ETag in `If-Match` was issued.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Precondition Failed
            status: 412
            detail: resource was modified, reload it and retry
            instance: /api/v1/calendars/0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10
            request_id: tQ6DNZRbQcyrEPPA
    UnsupportedMediaType:
      description: The request body has the wrong `Content-Type`.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Unsupported Media Type
            status: 415
            detail: content type must be application/json
            instance: /api/v1/calendars
            request_id: tQ6DNZRbQcyrEPPA
    UnprocessableEntity:
      description: The request body is well formed but some fields are invalid.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Unprocessable Entity
            status: 422
            detail: One or more fields are invalid.
            instance: /api/v1/calendars
            request_id: tQ6DNZRbQcyrEPPA
            errors:
              - field: color
                message: "must be a hex color like #3b6ea5"
    TooManyRequests:
      description: The rate limit was exceeded.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Too Many Requests
            status: 429
            detail: too many attempts, retry in 12s
            instance: /api/v1/me
            request_id: tQ6DNZRbQcyrEPPA
  schemas:
...
    Calendar:
      type: object
      description: A calendar that groups events.
      additionalProperties: false
      required: [id, name, color, description, timezone, kind, read_only, is_default, hidden, position, default_reminders, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          description: Calendar ID.
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name shown in the interface.
        color:
          type: string
          pattern: "^#[0-9a-f]{6}$"
          description: Color of the calendar's events as a lowercase hex value.
        description:
          type: string
          maxLength: 2000
          description: Free text description. Empty when not set.
        timezone:
          type: string
          nullable: true
          description: IANA timezone for new events, or null to use the account timezone.
        kind:
          type: string
          enum: [local, subscription]
          description: "`local` for calendars kept in this app, `subscription` for calendars filled from an ICS feed."
        read_only:
          type: boolean
          description: Whether events of this calendar can be changed. Subscription calendars are read only.
        is_default:
          type: boolean
          description: Whether new events go here when no calendar is chosen. The default calendar cannot be deleted.
        hidden:
          type: boolean
          description: Whether the calendar is hidden from calendar views.
        position:
          type: integer
          minimum: 0
          description: Sort order in lists, lowest first.
        default_reminders:
          type: array
          maxItems: 5
          description: Minutes before start at which new events of this calendar remind by default.
          items:
            type: integer
            minimum: 0
            maximum: 40320
        created_at:
          type: string
          format: date-time
          description: When the calendar was created.
        updated_at:
          type: string
          format: date-time
          description: When the calendar was last changed.
    CalendarList:
      type: object
      description: All calendars of the account.
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          description: Calendars ordered by position.
          items:
            $ref: "#/components/schemas/Calendar"
    CalendarInput:
      type: object
      description: >-
        Calendar fields to set. On create only `name` is required. On update send only the fields to change.
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name shown in the interface.
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          description: Hex color. Stored in lowercase.
        description:
          type: string
          maxLength: 2000
          description: Free text description.
        timezone:
          type: string
          nullable: true
          description: IANA timezone for new events. Null or an empty string uses the account timezone.
        hidden:
          type: boolean
          description: Hide the calendar from calendar views.
        position:
          type: integer
          minimum: 0
          maximum: 10000
          description: Sort order in lists, lowest first.
        default_reminders:
          type: array
          maxItems: 5
          description: Minutes before start for default reminders. Duplicates are removed.
          items:
            type: integer
            minimum: 0
            maximum: 40320
    Problem:
      type: object
      description: Error details in the RFC 9457 format.
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          description: URI that identifies the error type. `about:blank` means the HTTP status explains it.
        title:
          type: string
          description: Short summary of the HTTP status.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Explanation specific to this request.
        instance:
          type: string
          description: Path of the request that failed.
        request_id:
          type: string
          description: Identifier of the request, the same as the `X-Request-ID` header.
        errors:
          type: array
          description: Invalid fields, present on `422` responses.
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      description: One invalid input field.
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
          description: Name of the field in the request.
        message:
          type: string
          description: What is wrong with the value.
...